GlidenUI: Move text manipulation functions to util.cpp
