+++ projects/msvc/GLideNUI.vcxproj
@@ -95,6 +95,7 @@
     <ClCompile Include="..\..\src\GLideNUI\GLideNUI.cpp" />
     <ClCompile Include="..\..\src\GLideNUI\ScreenShot.cpp" />
     <ClCompile Include="..\..\src\GLideNUI\Settings.cpp" />
+    <ClCompile Include="..\..\src\GLideNUI\util.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\GLideNUI\About.h" />
@@ -103,6 +104,7 @@
     <ClInclude Include="..\..\src\GLideNUI\FullscreenResolutions.h" />
     <ClInclude Include="..\..\src\GLideNUI\GlSettings.h" />
     <ClInclude Include="..\..\src\GLideNUI\IniFileClass.h" />
+    <ClInclude Include="..\..\src\GLideNUI\util.h" />
     <ClInclude Include="..\..\src\GLideNUI\WTL\atlapp.h" />
     <ClInclude Include="..\..\src\GLideNUI\WTL\atlcrack.h" />
     <ClInclude Include="..\..\src\GLideNUI\WTL\atlctrls.h" />
