+++ src/GLideNUI/wtl.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#undef OS_WINDOWS
 #pragma warning(push)
 #pragma warning(disable : 4091) // warning C4091: 'typedef ': ignored on left of 'tagGPFIDL_FLAGS' when no variable is declared
 #pragma warning(disable : 4201) // warning C4201: nonstandard extension used: nameless struct/union
@@ -9,10 +10,10 @@
 #pragma warning(disable : 4996) // warning C4996: 'GetVersionExA': was declared deprecated
 #define _ATL_DISABLE_NOTHROW_NEW
 #include <atlbase.h>
-#undef DrawText
 #include "wtl\atlapp.h"
 #include "wtl\atlcrack.h"
 #include "wtl\atlctrls.h"
 #include "WTL\atlmisc.h"
+#include "WTL\atlctrlx.h"
 #include <atlwin.h>
 #pragma warning(pop)
